+++ changelog.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+##[2.1.0] - 2017-11-5
+##Changed
+- Task.php to reflect getter functions
+
+##Added
+- phpunit.xml.dist to root of project
+- Tasktest.php class to models folder w/ 3 validations
+
 ##[2.0.0] - 2017-11-2
 ### Added
 - Entity class to `/application/core/`